Rewrite the function to retrieve questions based on subject level and group number
