--- server/models/Questions.js
@@ -1,41 +1,0 @@
-const db = require("../db/connect")
-class Questions{
-    constructor({ question, option_1, option_2, option_3, option_4, answer, subject, level, group, created_at, updated_at}) {
-        this.question = question;
-        this.option_1 = option_1;
-        this.option_2 = option_2;
-        this.option_3 = option_3
-        this.option_4 = option_4
-        this.answer = answer
-        this.subject = subject
-        this.level = level
-        this.group = group
-        this.created_at = created_at
-        this.updated_at = updated_at
-    }
-static async getAll(){
-    try{
-        const response = await db.query("SELECT * FROM question_bank")
-        if(response.rows.length <0){
-            throw new Error("DB does not contain any data")
-        }
-        return response.rows.map(c => new Questions)
-    }
-    catch(error){
-        return error
-    }
-}
-static async getOneByGroup(id){
-    try{
-        const response = await db.query("SELECT * FROM question_bank WHERE group = $1",[id])
-        if(response.rows.length <0){
-            throw new Error("No Questions in that group")
-        }
-        return response.rows.map(c => new Questions)
-    }
-    catch(error){
-        return error
-    }
-}
-}
-module.exports = Questions
